refactor: NumberOfGuests 패키지 변경 -> vo로 이동
