+++ src/main/java/kitchenpos/common/vo/Quantity.java
@@ -10,7 +10,6 @@ public class Quantity {
     private long quantity;
 
     protected Quantity() {
-
     }
 
     public Quantity(long quantity) {
