+++ src/test/java/kitchenpos/ordertable/domain/TableGroupTest.java
@@ -11,6 +11,7 @@ import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.ordertable.exception.GroupTablesException;
 import kitchenpos.ordertable.exception.UngroupTablesException;
+import kitchenpos.ordertable.vo.NumberOfGuests;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
