+++ src/test/java/kitchenpos/ordertable/ui/TableRestControllerTest.java
@@ -14,11 +14,11 @@ import java.util.Arrays;
 import java.util.List;
 import kitchenpos.common.CommonTestFixtures;
 import kitchenpos.ordertable.application.TableService;
-import kitchenpos.ordertable.domain.NumberOfGuests;
 import kitchenpos.ordertable.domain.OrderTable;
 import kitchenpos.ordertable.dto.OrderTableRequest;
 import kitchenpos.ordertable.dto.OrderTableResponse;
 import kitchenpos.ordertable.testfixtures.TableTestFixtures;
+import kitchenpos.ordertable.vo.NumberOfGuests;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
